requirements.txt
langchain
langchain-core
sentence-transformers
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn